requirements.txt
Flask==2.0.3
Flask-SocketIO==5.2.0
paho-mqtt==1.6.1
eventlet==0.33.0
numpy==1.23.5
pandas==2.2.3
scikit-learn==1.0.2
joblib==1.1.0
Werkzeug==2.0.3
dnspython==1.16.0
python-dotenv==0.21.0
mpu6050-raspberrypi==1.2
scipy==1.15.2
smbus==1.1.post2
threadpoolctl==3.6.0